+++ javaparser-core-testing/src/test/java/com/github/javaparser/ast/type/TypeTest.java
@@ -26,7 +26,7 @@ import com.github.javaparser.ParseProblemException;
 import com.github.javaparser.ParseResult;
 import com.github.javaparser.ParserConfiguration;
 import com.github.javaparser.ast.expr.VariableDeclarationExpr;
-import com.github.javaparser.ast.validator.Java5Validator;
+import com.github.javaparser.ast.validator.language_level_validations.Java5Validator;
 import org.junit.jupiter.api.Test;
 
 import static com.github.javaparser.ParseStart.VARIABLE_DECLARATION_EXPR;
