--- javaparser-core/src/main/java/com/github/javaparser/ast/validator/Java1_3Validator.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (C) 2007-2010 Júlio Vilmar Gesser.
- * Copyright (C) 2011, 2013-2020 The JavaParser Team.
- *
- * This file is part of JavaParser.
- *
- * JavaParser can be used either under the terms of
- * a) the GNU Lesser General Public License as published by
- *     the Free Software Foundation, either version 3 of the License, or
- *     (at your option) any later version.
- * b) the terms of the Apache License
- *
- * You should have received a copy of both licenses in LICENCE.LGPL and
- * LICENCE.APACHE. Please refer to those files for details.
- *
- * JavaParser is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- */
-
-package com.github.javaparser.ast.validator;
-
-/**
- * This validator validates according to Java 1.3 syntax rules.
- */
-public class Java1_3Validator extends Java1_2Validator {
-    public Java1_3Validator() {
-        super();
-    }
-}
