+++ javaparser-core/src/main/java/com/github/javaparser/ast/validator/language_level_validations/Java12PreviewValidator.java
@@ -19,20 +19,28 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.ast.validator;
+package com.github.javaparser.ast.validator.language_level_validations;
 
 /**
- * This validator validates according to Java 13 syntax rules.
+ * This validator validates according to Java 12 syntax rules -- including incubator, preview, and second preview features.
  *
- * @see <a href="https://openjdk.java.net/projects/jdk/13/">https://openjdk.java.net/projects/jdk/13/</a>
+ * @see <a href="https://openjdk.java.net/projects/jdk/12/">https://openjdk.java.net/projects/jdk/12/</a>
  */
-public class Java13Validator extends Java12Validator {
+public class Java12PreviewValidator extends Java12Validator {
 
-    public Java13Validator() {
+    public Java12PreviewValidator() {
         super();
 
-        // Released Language Features
-        // No new released language features added in Java 13
+        // Incubator
+        // No new incubator language features added in Java 12
+
+        // Preview
+        remove(noSwitchExpressions); // Switch Expressions - first preview in Java 12 - https://openjdk.java.net/jeps/325
+        remove(onlyOneLabelInSwitchCase); // Switch Expressions - first preview in Java 12 - https://openjdk.java.net/jeps/325
+
+        // 2nd Preview
+        // No new 2nd preview language features added in Java 12
+
 
     }
 }
