+++ javaparser-core/src/main/java/com/github/javaparser/ast/validator/language_level_validations/Java12Validator.java
@@ -19,20 +19,20 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.ast.validator;
+package com.github.javaparser.ast.validator.language_level_validations;
 
 /**
- * This validator validates according to Java 14 syntax rules.
+ * This validator validates according to Java 12 syntax rules.
  *
- * @see <a href="https://openjdk.java.net/projects/jdk/14/">https://openjdk.java.net/projects/jdk/14/</a>
+ * @see <a href="https://openjdk.java.net/projects/jdk/12/">https://openjdk.java.net/projects/jdk/12/</a>
  */
-public class Java14Validator extends Java13Validator {
+public class Java12Validator extends Java11Validator {
 
-    public Java14Validator() {
+    public Java12Validator() {
         super();
 
         // Released Language Features
-        remove(noSwitchExpressions); // Switch Expressions - released in Java 14 - https://openjdk.java.net/jeps/361
-        remove(noYield); // Switch Expressions - released in Java 14 - https://openjdk.java.net/jeps/361
+        // No new released language features added in Java 14
+
     }
 }
