+++ javaparser-core/src/main/java/com/github/javaparser/ast/validator/language_level_validations/Java8Validator.java
@@ -19,10 +19,12 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.ast.validator;
+package com.github.javaparser.ast.validator.language_level_validations;
 
 import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
-import com.github.javaparser.ast.validator.chunks.ModifierValidator;
+import com.github.javaparser.ast.validator.SingleNodeTypeValidator;
+import com.github.javaparser.ast.validator.Validator;
+import com.github.javaparser.ast.validator.language_level_validations.chunks.ModifierValidator;
 
 /**
  * This validator validates according to Java 8 syntax rules.
