+++ javaparser-core/src/main/java/com/github/javaparser/ast/validator/language_level_validations/Java9Validator.java
@@ -19,11 +19,13 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.ast.validator;
+package com.github.javaparser.ast.validator.language_level_validations;
 
 import com.github.javaparser.ast.stmt.TryStmt;
-import com.github.javaparser.ast.validator.chunks.ModifierValidator;
-import com.github.javaparser.ast.validator.chunks.UnderscoreKeywordValidator;
+import com.github.javaparser.ast.validator.SingleNodeTypeValidator;
+import com.github.javaparser.ast.validator.Validator;
+import com.github.javaparser.ast.validator.language_level_validations.chunks.ModifierValidator;
+import com.github.javaparser.ast.validator.language_level_validations.chunks.UnderscoreKeywordValidator;
 
 /**
  * This validator validates according to Java 9 syntax rules.
