+++ javaparser-core/src/main/java/com/github/javaparser/ast/validator/language_level_validations/chunks/ModifierValidator.java
@@ -19,7 +19,7 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.ast.validator.chunks;
+package com.github.javaparser.ast.validator.language_level_validations.chunks;
 
 import com.github.javaparser.ast.Modifier;
 import com.github.javaparser.ast.body.*;
