+++ javaparser-core/src/main/java/com/github/javaparser/ast/validator/language_level_validations/chunks/NoUnderscoresInIntegerLiteralsValidator.java
@@ -19,30 +19,30 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.ast.validator.chunks;
+package com.github.javaparser.ast.validator.language_level_validations.chunks;
 
-import com.github.javaparser.ast.Node;
-import com.github.javaparser.ast.expr.Name;
-import com.github.javaparser.ast.expr.SimpleName;
+import com.github.javaparser.ast.expr.IntegerLiteralExpr;
+import com.github.javaparser.ast.expr.LiteralStringValueExpr;
+import com.github.javaparser.ast.expr.LongLiteralExpr;
 import com.github.javaparser.ast.validator.ProblemReporter;
 import com.github.javaparser.ast.validator.VisitorValidator;
 
-public class UnderscoreKeywordValidator extends VisitorValidator {
+public class NoUnderscoresInIntegerLiteralsValidator extends VisitorValidator {
     @Override
-    public void visit(Name n, ProblemReporter arg) {
-        validateIdentifier(n, n.getIdentifier(), arg);
+    public void visit(IntegerLiteralExpr n, ProblemReporter arg) {
+        validate(n, arg);
         super.visit(n, arg);
     }
 
     @Override
-    public void visit(SimpleName n, ProblemReporter arg) {
-        validateIdentifier(n, n.getIdentifier(), arg);
+    public void visit(LongLiteralExpr n, ProblemReporter arg) {
+        validate(n, arg);
         super.visit(n, arg);
     }
 
-    private static void validateIdentifier(Node n, String id, ProblemReporter arg) {
-        if (id.equals("_")) {
-            arg.report(n, "'_' is a reserved keyword.");
+    private static void validate(LiteralStringValueExpr n, ProblemReporter arg) {
+        if (n.getValue().contains("_")) {
+            arg.report(n, "Underscores in literal values are not supported.");
         }
     }
 }
