+++ javaparser-core/src/main/java/com/github/javaparser/ast/validator/language_level_validations/chunks/VarValidator.java
@@ -19,7 +19,7 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.ast.validator.chunks;
+package com.github.javaparser.ast.validator.language_level_validations.chunks;
 
 import com.github.javaparser.ast.Node;
 import com.github.javaparser.ast.body.Parameter;
